requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
PyYAML==5.4.1
urllib3==1.26.6

